+++ src/app/calorie-calculator/calorie-summary/calories.model.ts
@@ -3,3 +3,5 @@ export interface CaloriesSummary {
     carbohydrate: number;
     fat: number;
 }
+
+export type CalorieKeys = keyof CaloriesSummary;
